moved filters to separate file
